+++ Lab2/lab2/src/main/java/org/example/SequentialMatrixMultiplier.java
@@ -2,7 +2,7 @@ package org.example;
 
 public class SequentialMatrixMultiplier implements IMatrixMultiplier {
     @Override
-    public Result Multiply(float[][] matrixA, float[][] matrixB) {
+    public Result Multiply(int[][] matrixA, int[][] matrixB) {
         var resultMatrix = MatrixFunctions.Multiply(matrixA, matrixB);
 
         return new Result(resultMatrix);
